+++ src/PowerPlantChallenge.WebApi/Models/ProductionPlan.cs
@@ -4,14 +4,14 @@ namespace PowerPlantChallenge.WebApi.Models
 {
     public class ProductionPlan
     {
-        public ProductionPlan(int load, Fuels fuels, IEnumerable<PowerPlant> powerPlants)
+        public ProductionPlan(double load, Fuels fuels, IEnumerable<PowerPlant> powerPlants)
         {
             Load = load;
             Fuels = fuels;
             PowerPlants = powerPlants;
         }
 
-        public int Load { get; }
+        public double Load { get; }
         public Fuels Fuels { get; }
         public IEnumerable<PowerPlant> PowerPlants { get; }
     }
